requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
dnspython==1.16.0
pymongo==3.12.0
python-telegram-bot==13.7
pytz==2021.3
six==1.16.0
tornado==6.1
tzlocal==3.0
